Only restart NRPE at the end of the chef-run
On systems that have systemd, the rate limiting will kick in and mark the nrpe service as failed if a lot of checks are changed. This causes the chef-run to fail.

Also, it makes sense to only restart NRPE once at the end of the run when all the files are in place.